% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path}
\alias{get_path}
\title{Reformat file path in clipboard}
\usage{
get_path(to_clipboard = T)
}
\arguments{
\item{to_clipboard}{Logical. If TRUE, the function will write the new path to the clipboard. If FALSE, it will return the new path.}
}
\value{
If to_clipboard is FALSE, the function returns the new path. If TRUE, it writes the new path to the clipboard.
}
\description{
By default windows pastes file paths with backslashes, which is inconsistent
with R's use of forward slashes. This function will reformat file paths copied to clipboard to be compatible with R.
}
\examples{
\dontrun{
# Can copy a file path in windows explorer using Shift + Right Click and selecting "Copy as path"
get_path() # reformat the path in the clipboard, can then use Ctrl + V to paste where needed
get_path(F) # pastes the new path to the console}
}
