% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_table.R
\name{print.r_tableObject}
\alias{print.r_tableObject}
\title{Print method for r_tableObject}
\usage{
\method{print}{r_tableObject}(x, ...)
}
\arguments{
\item{x}{An object of class \code{r_tableObject}.}

\item{...}{Further arguments passed to or from other methods (not used by this method).}
}
\value{
Invisibly returns the original object \code{x}.
}
\description{
Custom print output for objects of class \code{r_tableObject}.
(Users typically do not call this directly, but via \code{print(object)}).
}
