% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlist.R
\name{varlist}
\alias{varlist}
\title{Search a dataframe for variable names matching a pattern}
\usage{
varlist(
  pattern = NULL,
  df = NULL,
  exclude = NULL,
  type = c("numeric", "factor", "character", "double", "logical", "integer"),
  ignore.case = TRUE
)
}
\arguments{
\item{pattern}{character string in regular expression format.}

\item{df}{dataframe object, if left as NULL it assumes the dataframe is labeled "df".}

\item{exclude}{character string in regular expression format identifying variable names to exclude. Defaults to NULL.}

\item{type}{character vector of variable types to include. Defaults to all types. Valid types are: "numeric", "factor", "character", "double", "logical", "integer".}

\item{ignore.case}{logical indicating whether to ignore case when matching the pattern. Defaults to TRUE.}
}
\value{
character vector of variable names found within selected dataframe.
}
\description{
Search a dataframe for variable names matching a pattern
}
\examples{
varlist("hsns", exclude="5|oj")

varlist(type="factor",exclude=c("gb|^c"))
varlist(pattern="neo[1-9]", df = data)
sapply(df[,varlist(type="numeric",pattern="credit")], summary)
}
\seealso{
\code{\link[=grepl]{grepl()}}
}
